refactor: Seperate test codes
- 통합 테스트 코드 기능별로 분리
