+++ src/test/java/com/wanted/preonboarding/BackendPreonboardingApplicationTests.java
@@ -1,9 +1,11 @@
 package com.wanted.preonboarding;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
+@AutoConfigureMockMvc
 class BackendPreonboardingApplicationTests {
 
     @Test
